Add ui.json and fix up params classes
