+++ grid_apps-assets/uijson/block_model_creation.ui.json
@@ -1,0 +1,92 @@
+{
+    "version": "0.1.0",
+    "title": "Block Model Creation",
+    "conda_environment": "grid_apps",
+    "run_command": "grid_apps.block_models.driver",
+    "geoh5": "",
+    "monitoring_directory": "",
+    "workspace_geoh5":  "",
+    "conda_environment_boolean": false,
+    "objects": {
+        "main": true,
+        "group": "Data",
+        "meshType": [
+            "{2e814779-c35f-4da0-ad6a-39a6912361f9}",
+            "{202c5db1-a56d-4004-9cad-baafd8899406}",
+            "{6a057fdc-b355-11e3-95be-fd84a7ffcb88}",
+            "{f26feba3-aded-494b-b9e9-b2bbcbe298e1}",
+            "{4ea87376-3ece-438b-bf12-3479733ded46}",
+            "{48f5054a-1c5c-4ca4-9048-80f36dc60a06}"
+        ],
+        "label": "Object",
+        "value": ""
+    },
+    "new_grid": {
+        "main": true,
+        "group": "Data",
+        "label": "Name",
+        "value": ""
+    },
+    "cell_size_x": {
+        "main": true,
+        "group": "Grid",
+        "label": "Minimum x cell size",
+        "value": 1.0,
+        "min": 1e-14
+    },
+    "cell_size_y": {
+        "main": true,
+        "group": "Grid",
+        "label": "Minimum y cell size",
+        "value": 1.0,
+        "min": 1e-14
+    },
+    "cell_size_z": {
+        "main": true,
+        "group": "Grid",
+        "label": "Minimum z cell size",
+        "value": 1.0,
+        "min": 1e-14
+    },
+    "depth_core": {
+        "main": true,
+        "group": "Grid",
+        "label": "Core depth (m)",
+        "value": 0.0,
+        "min": 0.0
+    },
+    "horizontal_padding": {
+        "main": true,
+        "group": "Grid",
+        "label": "Horizontal padding",
+        "value": 0.0,
+        "min": 0.0
+    },
+    "bottom_padding": {
+        "main": true,
+        "group": "Grid",
+        "label": "Bottom padding",
+        "value": 0.0,
+        "min": 0.0
+    },
+    "expansion_fact": {
+        "main": true,
+        "group": "Grid",
+        "label": "Expansion factor",
+        "value": 0.0
+    },
+    "export_as": {
+        "group": "Output",
+        "main": true,
+        "label": "Save as",
+        "value": "block_model"
+    },
+    "out_group": {
+        "group": "Output",
+        "main": true,
+        "optional": true,
+        "enabled": false,
+        "label": "Group",
+        "value": ""
+    }
+}
