--- grid_apps-assets/uijson/hello.ui.json
@@ -1,11 +1,0 @@
-{
-    "title": "grid-apps Hello World",
-    "conda_environment": "grid_apps",
-    "run_command": "grid_apps.commands.hello",
-    "name": {
-        "main": true,
-        "label": "Name",
-        "value": "World",
-        "tooltip": "Name to say hello to"
-    }
-}
