+++ grid_apps/__init__.py
@@ -8,10 +8,11 @@
 
 from __future__ import annotations
 
+import logging
 from pathlib import Path
 
-
 __version__ = "0.1.0-alpha.1"
+logging.basicConfig(level=logging.INFO)
 
 
 def assets_path() -> Path:
