+++ grid_apps/block_models/__init__.py
@@ -1,10 +1,7 @@
 # ''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
-#  Copyright (c) 2022-2024 Mira Geoscience Ltd.                                '
+#  Copyright (c) 2024 Mira Geoscience Ltd.                                     '
 #                                                                              '
-#  This file is part of grid-apps package.                                        '
+#  This file is part of grid-apps package.                                     '
 #                                                                              '
 #  All rights reserved.                                                        '
 # ''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
-
-if __name__ == "__main__":  # pragma: no cover
-    print("Hello, world!")
